requirements.txt
asgiref==3.2.10
Django==3.0.8
djangorestframework==3.11.0
python-decouple==3.3
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
Pillow==7.2.0
drf-yasg==1.17.1
geocoder==1.38.1
geopy==2.0.0